==> analysis/02.entry_point.ipynb <==
# MERFISH Experiment Region Entry Point

## Goals
- Provide a single entry point for a MERFISH experiment region
- Simple API for retriving file and experimental information
- Plot MERFISH transcripts together with DAPI, PolyT and other smFISH images

## Main Entry Point

from merfishing import Merfish
import pandas as pd
import matplotlib.pyplot as plt

merfish = Merfish('./output/region_0')

## Get Info

# field of view id
fov =int(merfish.fov_ids[100])
# padding in pixel to add to the image boarder
padding = 150

# get FOV cell meta
cell_meta = merfish.get_cell_metadata(fov=fov)

# get FOV cell boundaries
boundaries = merfish.get_cell_boundaries(fov=fov)

# get FOV transcripts
transcripts = merfish.get_transcripts(fov)

## Transform Coordinates

### Four kinds of coordinates and their transformation 
- Global Pixel
- Global Micron
- Local Pixel
- Local Micron

### Transform Helper
- get FOV extent: `xmin, ymin, xmax, ymax = merfish.get_fov_pixel_extent_from_transcripts(fov, padding)`
- get FOV offset: `offset = (xmin, ymin)`
- Global to Local: add FOV offset (xmin, ymin)
- Local to Global: minus FOV offset (xmin, ymin)
- Micron to Pixel: `merfish.transform.micron_to_pixel_transform`
- Pixel to Micron: `merfish.transform.pixel_to_micron_transform`

# coords in file are global microns
cell_meta[['center_x', 'center_y']].head()

merfish.transform.micron_to_pixel_transform([[280, 2399], [306, 2501]])

merfish.transform.pixel_to_micron_transform([[1720, 24564], [1960, 25508]])

# xmin, ymin, xmax, ymax
merfish.get_fov_micron_extent_from_transcripts(fov)

# xmin, ymin, xmax, ymax
merfish.get_fov_pixel_extent_from_transcripts(fov, padding)

## Make Plots

## More control on images

### Plot transcripts

transcripts['gene'].value_counts().index[:50]

genes = transcripts['gene'].value_counts().index[:50]

fig = merfish.plot_fov(fov=fov, genes=genes, dpi=150)

### plot other smFISH Images

fig, fov_images = merfish.plot_fov(
    fov=fov,
    plot_boundary=False,
    plot_cell_centers=False,
    genes=None,
    image_names=('PolyT+Snap25+DAPI', 'DAPI', 'PolyT', 'Snap25', 'Mbp'),
    padding=padding,
    dpi=150
)

fig, fov_images = merfish.plot_fov(
    fov=fov,
    plot_boundary=True,
    plot_cell_centers=False,
    genes=None,
    image_names=('PolyT+Snap25+DAPI', 'DAPI', 'PolyT', 'Snap25', 'Mbp'),
    padding=padding,
    dpi=150,
    filter = True,
)
