==> analysis/03.detect_smfish_spots.ipynb <==
from merfishing import Merfish

## Entry Point

merfish = Merfish('/home/qz/hanliu/cemba3c/merfish/202208121423_MouseSagittalM3S2_VMSC01101/output/region_0')

## Call Spots

merfish.smfish_genes

merfish.call_spots(
    image_names='all',
    cpu=15,
    padding=50,
    detect_dense=True,
    projection='max',
    verbose=False,
    redo=False,
)

## Plot

fig = merfish.plot_fov(
    0,
    n_cols=3,
    image_names=['PolyT+Snap25+DAPI', 'PolyT', 'DAPI'] + merfish.smfish_genes, 
    genes=merfish.smfish_genes,
    dpi=150
)

## Transcripts

transcripts = merfish.get_transcripts(fov=0)
transcripts[transcripts['gene'].isin(merfish.smfish_genes)]

## Details - BigFish Package

Spot detection is done by the bigfish package with its automatic pipeline and default parameters. Each z-plane is run separately. Please read the bigfish package for more details and fine tune the parameter if needed. You can pass the key parameters to merfish.call_spots function.
- Github: https://github.com/fish-quant/big-fish
- Documentation: https://big-fish.readthedocs.io/en/stable/index.html
- Jupyter Notebook Example: https://github.com/fish-quant/big-fish-examples/tree/master/notebooks

from bigfish.stack import maximum_projection
from bigfish.plot import plot_detection, plot_images
from bigfish.detection import detect_spots

image = merfish.get_image_fov('Snap25', 2)
plot_images(maximum_projection(image), contrast=True, framesize=(3, 3))

spots, threshold = detect_spots(
    images=image[3], 
    return_threshold=True, 
    voxel_size=(103, 103),  # in nanometer (one value per dimension zyx)
    spot_radius=(150, 150))  # in nanometer (one value per dimension zyx)
print("detected spots")
print("\r shape: {0}".format(spots.shape))
print("\r dtype: {0}".format(spots.dtype))
print("\r threshold: {0}".format(threshold))

plot_detection(image[3], spots, contrast=True, framesize=(6, 6), radius=1)
